+++ addon/components/time-picker.js
@@ -1,9 +1,9 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import moment from 'moment';
-import parseTime from 'ember-date-components/utils/parse-time';
-import buildTimeRange from 'ember-date-components/utils/build-time-range';
-import { shouldUseAmPm } from 'ember-date-components/utils/should-use-am-pm';
+import parseTime from '@gorner/ember-date-components/utils/parse-time';
+import buildTimeRange from '@gorner/ember-date-components/utils/build-time-range';
+import { shouldUseAmPm } from '@gorner/ember-date-components/utils/should-use-am-pm';
 import { guidFor } from '@ember/object/internals';
 import { tracked } from '@glimmer/tracking';
 import { assert } from '@ember/debug';