+++ app/components/date-picker-month-year-select.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-date-components/components/date-picker-month-year-select';
+export { default } from '@gorner/ember-date-components/components/date-picker-month-year-select';