+++ app/components/date-picker-month.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-date-components/components/date-picker-month';
+export { default } from '@gorner/ember-date-components/components/date-picker-month';