+++ app/components/date-picker-navigation.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-date-components/components/date-picker-navigation';
+export { default } from '@gorner/ember-date-components/components/date-picker-navigation';