+++ app/helpers/date-picker-day-classes.js
@@ -1,4 +1,4 @@
 export {
   default,
   datePickerDayClasses,
-} from 'ember-date-components/helpers/date-picker-day-classes';
+} from '@gorner/ember-date-components/helpers/date-picker-day-classes';