+++ app/helpers/is-equal.js
@@ -1,1 +1,1 @@
-export { default, isEqual } from 'ember-date-components/helpers/is-equal';
+export { default, isEqual } from '@gorner/ember-date-components/helpers/is-equal';